Improved Cast processing

+++ onnx2tf/ops/Gather.py
@@ -75,7 +75,7 @@ def make_node(
     # tensorflow gather supports only positive indices
     if not optimization_for_gpu_delegate:
         cond = tf.cast(indices < 0, dtype=indices.dtype)
-        indices = indices + tf.cast(cond * tf.shape(input_tensor)[axis], dtype=indices.dtype)
+        indices = tf.cast(indices + cond * tf.shape(input_tensor)[axis], dtype=tf.int64)
     else:
         cond = indices < 0
         indices = indices + tf.cast(tf.where(cond, 1, 0) * tf.shape(input_tensor)[axis], dtype=indices.dtype)
